added simple layout implementation for explorer plugin
